Refactor predicate test files
Move predicate test files from
package `predicate` to package `predicates`
