+++ src/test/java/seedu/realodex/model/person/predicates/NameContainsKeyphrasePredicateTest.java
@@ -1,4 +1,4 @@
-package seedu.realodex.model.person.predicate;
+package seedu.realodex.model.person.predicates;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -7,7 +7,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 
 import seedu.realodex.model.person.Person;
-import seedu.realodex.model.person.predicates.NameContainsKeyphrasePredicate;
 import seedu.realodex.testutil.PersonBuilder;
 
 public class NameContainsKeyphrasePredicateTest {
