+++ src/test/java/seedu/realodex/model/person/predicates/TagsMatchPredicateTest.java
@@ -1,4 +1,4 @@
-package seedu.realodex.model.person.predicate;
+package seedu.realodex.model.person.predicates;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -9,7 +9,6 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 
 import seedu.realodex.model.person.Person;
-import seedu.realodex.model.person.predicates.TagsMatchPredicate;
 import seedu.realodex.model.tag.Tag;
 import seedu.realodex.testutil.PersonBuilder;
